Make FlyDSL mandatory on native ROCm builds and consolidate the package
FlyDSL was an opt-in extra on all builds. On native ROCm builds the variant
is known at build time, so inject flydsl into install_requires directly,
making it a hard requirement; python-only builds remain variant-agnostic and
keep it as an opt-in extra.

Move the pinned version into ci/flydsl_version.txt as a single source of
truth, read by both setup.py and ci/scripts/utils_flydsl.bash, and bump the
pin to 0.2.4.

Consolidate the FlyDSL support code under the mslk.flydsl package (follow-up
from #431): mslk/utils/flydsl.py and flydsl_aot.py move to mslk/flydsl/, split
into common.py (availability + shared cache path), jit.py (runtime dispatch),
and aot.py (pre-compilation). Tests mirror the split under test/flydsl/.
