+++ ci/flydsl_version.txt
@@ -1,0 +1,1 @@
+0.2.4
